refactor: props기반 라우팅 처리
